refactor: 게임을 실행하는 main 메소드에서 input, play, output 이 수행되도록 개선
